BC-1310 fix tests
- the tests with s3 comminication should into a file
